chore: migrate importNetworks function
