+++ axelar-chains-config/utils/index.js
@@ -1,2 +1,3 @@
 export * from './getChainArray';
 export * from './readJSON';
+export * from './importNetworks';
